+++ security/provenance/include/provenance.h
@@ -49,7 +49,7 @@ enum {
 };
 
 struct provenance {
-	union prov_msg msg;
+	union prov_elt msg;
 	spinlock_t lock;
 	uint8_t updt_mmap;
 	uint8_t has_mmap;
@@ -58,7 +58,7 @@ struct provenance {
 	bool saved;
 };
 
-#define prov_msg(provenance) (&(provenance->msg))
+#define prov_elt(provenance) (&(provenance->msg))
 #define prov_lock(provenance) (&(provenance->lock))
 
 #define ASSIGN_NODE_ID 0
@@ -72,10 +72,10 @@ static inline struct provenance *alloc_provenance(uint64_t ntype, gfp_t gfp)
 	if (!prov)
 		return NULL;
 	spin_lock_init(prov_lock(prov));
-	prov_type(prov_msg(prov)) = ntype;
-	node_identifier(prov_msg(prov)).id = prov_next_node_id();
-	node_identifier(prov_msg(prov)).boot_id = prov_boot_id;
-	node_identifier(prov_msg(prov)).machine_id = prov_machine_id;
+	prov_type(prov_elt(prov)) = ntype;
+	node_identifier(prov_elt(prov)).id = prov_next_node_id();
+	node_identifier(prov_elt(prov)).boot_id = prov_boot_id;
+	node_identifier(prov_elt(prov)).machine_id = prov_machine_id;
 	return prov;
 }
 
@@ -89,7 +89,7 @@ static inline void copy_node_info(union prov_identifier *dest, union prov_identi
 	memcpy(dest, src, sizeof(union prov_identifier));
 }
 
-static inline void __record_node(union prov_msg *node)
+static inline void __record_node(union prov_elt *node)
 {
 	if (filter_node(node) || provenance_is_recorded(node)) // filtered or already recorded
 		return;
@@ -103,7 +103,7 @@ static inline void __record_node(union prov_msg *node)
 static inline void __prepare_relation(uint64_t type,
 				      union prov_identifier *from,
 				      union prov_identifier *to,
-				      union prov_msg *relation,
+				      union prov_elt *relation,
 				      struct file *file)
 {
 	prov_type(relation) = type;
@@ -120,8 +120,8 @@ static inline void __prepare_relation(uint64_t type,
 
 static inline int __update_version(uint64_t type, struct provenance *prov)
 {
-	union prov_msg old_prov;
-	union prov_msg relation;
+	union prov_elt old_prov;
+	union prov_elt relation;
 	int rc = 0;
 
 	if (!prov->has_outgoing) // there is no outgoing
@@ -129,16 +129,16 @@ static inline int __update_version(uint64_t type, struct provenance *prov)
 	if (filter_update_node(type))
 		return 0;
 
-	memset(&relation, 0, sizeof(union prov_msg));
-	memcpy(&old_prov, prov_msg(prov), sizeof(union prov_msg));
-	memset(prov_taint(prov_msg(prov)), 0, PROV_N_BYTES);
-	node_identifier(prov_msg(prov)).version++;
-	clear_recorded(prov_msg(prov));
-	if (node_identifier(prov_msg(prov)).type == ACT_TASK)
-		__prepare_relation(RL_VERSION_PROCESS, &(old_prov.msg_info.identifier), &(prov_msg(prov)->msg_info.identifier), &relation, NULL);
+	memset(&relation, 0, sizeof(union prov_elt));
+	memcpy(&old_prov, prov_elt(prov), sizeof(union prov_elt));
+	memset(prov_taint(prov_elt(prov)), 0, PROV_N_BYTES);
+	node_identifier(prov_elt(prov)).version++;
+	clear_recorded(prov_elt(prov));
+	if (node_identifier(prov_elt(prov)).type == ACT_TASK)
+		__prepare_relation(RL_VERSION_PROCESS, &(old_prov.msg_info.identifier), &(prov_elt(prov)->msg_info.identifier), &relation, NULL);
 	else
-		__prepare_relation(RL_VERSION, &(old_prov.msg_info.identifier), &(prov_msg(prov)->msg_info.identifier), &relation, NULL);
-	rc = call_query_hooks(&old_prov, prov_msg(prov), &relation);
+		__prepare_relation(RL_VERSION, &(old_prov.msg_info.identifier), &(prov_elt(prov)->msg_info.identifier), &relation, NULL);
+	rc = call_query_hooks(&old_prov, prov_elt(prov), &relation);
 	prov_write(&relation);
 	prov->has_outgoing = false; // we update there is no more outgoing edge
 	prov->saved = false;
@@ -151,26 +151,26 @@ static inline int record_relation(uint64_t type,
 				  uint8_t allowed,
 				  struct file *file)
 {
-	union prov_msg relation;
+	union prov_elt relation;
 	int rc = 0;
 
 	// check if the nodes match some capture options
-	apply_target(prov_msg(from));
-	apply_target(prov_msg(to));
+	apply_target(prov_elt(from));
+	apply_target(prov_elt(to));
 
-	if (!provenance_is_tracked(prov_msg(from)) && !provenance_is_tracked(prov_msg(to)) && !prov_all)
+	if (!provenance_is_tracked(prov_elt(from)) && !provenance_is_tracked(prov_elt(to)) && !prov_all)
 		return 0;
-	if (!should_record_relation(type, prov_msg(from), prov_msg(to)))
+	if (!should_record_relation(type, prov_elt(from), prov_elt(to)))
 		return 0;
-	memset(&relation, 0, sizeof(union prov_msg));
-	__record_node(prov_msg(from));
-	__record_node(prov_msg(to));
+	memset(&relation, 0, sizeof(union prov_elt));
+	__record_node(prov_elt(from));
+	__record_node(prov_elt(to));
 	rc = __update_version(type, to);
 	if (rc < 0)
 		return rc;
-	__record_node(prov_msg(to));
-	__prepare_relation(type, &(prov_msg(from)->msg_info.identifier), &(prov_msg(to)->msg_info.identifier), &relation, file);
-	rc = call_query_hooks(prov_msg(from), prov_msg(to), &relation);
+	__record_node(prov_elt(to));
+	__prepare_relation(type, &(prov_elt(from)->msg_info.identifier), &(prov_elt(to)->msg_info.identifier), &relation, file);
+	rc = call_query_hooks(prov_elt(from), prov_elt(to), &relation);
 	prov_write(&relation);
 	from->has_outgoing = true; // there is an outgoing edge
 	return rc;
@@ -183,7 +183,7 @@ static inline void flow_to_activity(uint64_t type,
 				    struct file *file)
 {
 	record_relation(type, from, to, allowed, file);
-	if (should_record_relation(type, prov_msg(from), prov_msg(to)))
+	if (should_record_relation(type, prov_elt(from), prov_elt(to)))
 		to->updt_mmap = 1;
 }